requirements.txt
numpy
pyzmq
lz4
msgpack
torch